added parallelism - default compiled with 2 workers per operator

+++ Operators/InOut/ReadFromFile.hpp
@@ -87,7 +87,7 @@ protected:
 		return OperatorClass::INPUT;
 	}
 
-	ff::ff_node* node_operator(size_t parallelism) {
+	ff::ff_node* node_operator(int parallelism) {
 		if(parallelism==1){
 			return new ReadFromFileFFNode<Out>(func, filename);
 		}
