+++ Operators/Map.hpp
@@ -92,7 +92,7 @@ protected:
 	}
 
 
-	ff::ff_node* node_operator(size_t parallelism){
+	ff::ff_node* node_operator(int parallelism){
 		if(parallelism == 1){
 			return new UnaryMapFFNode<In, Out>(&mapf);
 		}
